Add files via upload
In this puzzle game, a random word is selected from the predefined list and presented as numbers. You have to guess the word from the number. You get points for every correct guess. When you reach the end of the game, it shows you a dialogue box and closes with your consent.